requirements.txt
torch>=2.0
open-clip-torch>=2.20.0
Pillow>=9.0
numpy>=1.24
fastapi>=0.100
uvicorn[standard]>=0.22
aiosqlite>=0.19
datasets>=2.14
